Some fixes to get at least partway through compilation...
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/array-rework@3233 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/LAPACK.hpp
@@ -210,9 +210,9 @@ extern "C" {
 
   void cgetri_(int *N, float *A, int *LDA, int *IPIV, float *WORK, int *LWORK, int *INFO);
 
-  void dgetri_(int *N, float *A, int *LDA, int *IPIV, float *WORK, int *LWORK, int *INFO);
+  void dgetri_(int *N, double *A, int *LDA, int *IPIV, double *WORK, int *LWORK, int *INFO);
   
-  void zgetri_(int *N, float *A, int *LDA, int *IPIV, float *WORK, int *LWORK, int *INFO);
+  void zgetri_(int *N, double *A, int *LDA, int *IPIV, double *WORK, int *LWORK, int *INFO);
 
   int xerbla_(char *srname, int *info);
